Update golang.org/x/mod
Newer go versions use v1.2.3 instead of v1.23. v0.6 of x/mod couldn't
parse the new versions, so it has been updated.

This also adds a go.sum file and lists indirect dependencies in the
go.mod, as `go mod tidy` was executed.

This change was not tested with older versions of go, however, x/mod
defines a minimum version of 1.18, therefore this is the new minimum
required for building.