Add support for Go 1.24 tools (#60)
go-mod-upgrade doesn't currently work with the new tool dependencies ([release notes for reference](https://go.dev/doc/go1.24#go-command)). 

The problem seems to be that go list doesn't natively handle -u -m for modules specified as tools yet. 
So this is just a simple work around that lists the tools and then finds the version to update one by one.